Replaced trend analysis with least squares fit of the MA instead of derivative of the MA.
Added function to rank stocks based on the gradient of the least squares fit. Using Monte Carlo analysis, the ranking is done N times with gradient values being taken as a random value from a normal distribution (mean = gradient, stddev = gradient error). The higher the gradient and the better the fit, the lower the ranking of the stock